All messages and suggestions are controlled via headers
`Messages` header
===============

This header is used to control the type of messages. It has
4 values: article, text, infographic, suggested-topics
One may use comma seperated list of values for testing

`Suggestions` header
====================

This header is used to control the tpe of suggestions. It has
4 values: Explore-Stabby, Explore-Questions, Suggest-Questions,
No-Suggest

Like the Messages header, one may use this header as comma separated
values too.